Update configuration body to issue's body
Since GitHub comment API do not support sorting, It takes more time to
get latest comment of issue.

So I decided to update configuration on issue's body and leaves comments
as a revision of a configuration.

`get-rev` command helps you to get the specific version of configuration
(it starts with index 0).